update the directory of database related feature
