+++ app/database/__init__.py
@@ -1,0 +1,17 @@
+# from flask import Flask
+# from flask_sqlalchemy import SQLAlchemy
+#
+# # Set db path
+# db_path = '../../data/song.sqlite'
+#
+# # initialized flask app
+# app = Flask(__name__)
+#
+# # SQLAlchemy configuration
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + db_path
+# app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
+#
+# # Initialize the database
+# db = SQLAlchemy(app)
+
+
