+++ app/database/create_db.py
@@ -1,0 +1,5 @@
+from schema import db
+
+if __name__=='__main__':
+    # create song.sqlite under data folder
+    db.create_all()
