+++ app/database/db_conn.py
@@ -1,0 +1,13 @@
+import sqlite3
+
+def dbConn(path):
+    """
+    conn to sqlite database
+
+    Returns
+    -------
+    sqlite database connection object
+    """
+    song_db = '../../data/song.sqlite'
+    conn = sqlite3.connect(path)
+    return conn
