--- app/msiapp/templates/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <link href="static/basic.css" rel="stylesheet">
-</head>
-
-<body>
-    <h3>
-         <a href = "{{ url_for('index') }}">Tracks I like</a>
-    </h3>
-
-    <hr/>
-    <table>
-         <thead>
-            <tr>
-               <th>Title</th>
-               <th>Artist</th>
-               <th>Album</th>
-            </tr>
-         </thead>
-
-         <tbody>
-            {% for track in tracks %}
-               <tr>
-                  <td>{{ track.title }}</td>
-                  <td>{{ track.artist }}</td>
-                  <td>{{ track.album }}</td>
-               </tr>
-            {% endfor %}
-         </tbody>
-      </table>
-
-</body>
-</html>
