--- develop/models/database/create_db.py
@@ -1,4 +1,0 @@
-from develop.models.database.schema import db
-
-# create song.sqlite under data folder
-db.create_all()
